near post work server

+++ Backend/src/post/group.go
@@ -10,8 +10,8 @@ func getGroupKey(uID uint64) string {
 	return "group:" + strconv.FormatUint(uID, 10)
 }
 
-func dbGetGroups(uID uint64, c redis.Conn) ([]int, error) {
-	var result []int
+func dbGetGroups(uID uint64, c redis.Conn) ([]uint32, error) {
+	var result []uint32
 
 	groupKey := getGroupKey(uID)
 	values, err := redis.Values(c.Do("SMEMBERS", groupKey))
@@ -20,17 +20,17 @@ func dbGetGroups(uID uint64, c redis.Conn) ([]int, error) {
 	}
 
 	for _, value := range values {
-		group, err := redis.Int(value, err)
+		group, err := redis.Uint64(value, err)
 		if err != nil {
 			return nil, err
 		}
-		result = append(result, group)
+		result = append(result, uint32(group))
 	}
 
 	return result, nil
 }
 
-func inSameGroup(groups1 []int, groups2 []int) bool {
+func inSameGroup(groups1 []uint32, groups2 []uint32) bool {
 	for _, group1 := range groups1 {
 		for _, group2 := range groups2 {
 			if group1 == group2 {
